+++ Kernel/System/DynamicField/Driver/Text.pm
@@ -301,7 +301,7 @@ sub SearchFieldParameterBuild {
     };
 }
 
-sub CommonSearchFieldParameterBuild {
+sub StatsSearchFieldParameterBuild {
     my ( $Self, %Param ) = @_;
 
     my $Value = $Param{Value};